more simulation stuff
