+++ simgui-ds.json
@@ -93,11 +93,12 @@
     {
       "guid": "Keyboard0"
     },
-    {
-      "guid": "Keyboard1"
-    },
+    {},
     {
       "guid": "Keyboard2"
+    },
+    {
+      "guid": "Keyboard1"
     }
   ]
 }
